WD-69: Moved the init classes under init package
