+++ Code/EnvironmentMetricsService/src/main/java/ie/tcd/wayfinder/envmetrics/init/ResourceServerConfig.java
@@ -1,0 +1,22 @@
+package ie.tcd.wayfinder.envmetrics.init;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
+import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
+
+//@Configuration
+//@EnableResourceServer
+//public class ResourceServerConfig extends ResourceServerConfigurerAdapter {
+//	@Override
+//	public void configure(HttpSecurity http) throws Exception {
+//		http
+//			.requestMatchers().antMatchers("/**")
+//		.and()
+//			.authorizeRequests()
+//				.antMatchers("/**").authenticated()
+//		.and()
+//			.cors().disable();
+//	}
+//	
+//}
