+++ Code/EnvironmentMetricsService/src/main/java/ie/tcd/wayfinder/envmetrics/init/WeatherController.java
@@ -1,13 +1,14 @@
-package ie.tcd.wayfinder.envmetrics.weather;
+package ie.tcd.wayfinder.envmetrics.init;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping(value="/")
 public class WeatherController {
 
-    @RequestMapping(value="/", method=RequestMethod.GET)
+    @RequestMapping(value="/abc", method=RequestMethod.GET)
     public String index()
     {
         return "Wayfinders Springboot Env Service!";
